server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oriself-server"
version = "0.1.0"
description = "OriSelf Next · FastAPI 后端。对话循环、guardrails、多 provider LLM。"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name = "niuniu-869"}]
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "sqlalchemy>=2.0",
    "httpx>=0.26",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
]
postgres = [
    "psycopg[binary]>=3.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["oriself_server*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
